Updated gold rain lane targeting
Wrong behavior: Currently checks ALL enemies if they are a worthy target.
Changes include: Do not consider an enemy for a gold rain target unless it is in the gold rain lane. Moved the for loop block of code into the if block of whether a gold rain ability is active.

Fixed whitespace

Was using the online editor :/

magic numbers are back